Add a CodeAnalysisMixin for shared methods
