+++ app/workers/commit_monitor_handlers/commit_range/rubocop_checker.rb
@@ -5,6 +5,7 @@ class CommitMonitorHandlers::CommitRange::RubocopChecker
   sidekiq_options :queue => :miq_bot_glacial
 
   include BranchWorkerMixin
+  include CodeAnalysisMixin
 
   def self.handled_branch_modes
     [:pr]
@@ -21,11 +22,7 @@ class CommitMonitorHandlers::CommitRange::RubocopChecker
   private
 
   def process_branch
-    unmerged_results = []
-    unmerged_results << Linter::Rubocop.new(branch).run
-    unmerged_results << Linter::Haml.new(branch).run
-    unmerged_results << Linter::Yaml.new(branch).run
-    unmerged_results.compact!
+    unmerged_results = run_all_linters
     if unmerged_results.empty?
       @results = {"files" => []}
     else
@@ -37,21 +34,6 @@ class CommitMonitorHandlers::CommitRange::RubocopChecker
     replace_rubocop_comments
   rescue GitService::UnmergeableError
     nil # Avoid working on unmergeable PRs
-  end
-
-  def merge_linter_results(*results)
-    return if results.empty?
-
-    new_results = results[0].dup
-
-    results[1..-1].each do |result|
-      %w(offense_count target_file_count inspected_file_count).each do |m|
-        new_results['summary'][m] += result['summary'][m]
-      end
-      new_results['files'] += result['files']
-    end
-
-    new_results
   end
 
   def rubocop_comments
